+++ CHANGELOG.md
@@ -2,8 +2,12 @@
 
 All notable changes to `eltabarani/reusable-query` will be documented in this file.
 
-## [1.0.0] - 2025-07-23
+## [1.1.0] - 2025-07-28
 
-### Added
+- Now the `useQuery` compatibility with Laravel's Eloquent Scopes and Closures
+- Added support for dynamic parameters in reusable queries
+- Improved README for clarity and examples
+
+## [1.0.0] - 2025-07-23
 
 - Initial stable release :)